requirements.txt
asgiref==3.7.2
Django==4.2.6
mysqlclient==2.2.0
Pillow==10.0.1
product==0.1.1.1
sqlparse==0.4.4
tzdata==2023.3
python-decouple==3.8